factor out image-related methods from util module
