+++ lib/app.js
@@ -10,6 +10,7 @@ var mkdirp = require('mkdirp');
 var rimraf = require('rimraf');
 var Progger = require('progger');
 var util = require(path.resolve(__dirname, './', 'util'));
+var image = require(path.resolve(__dirname, './', 'image'));
 var msg = require(path.resolve(__dirname, './', 'msg'));
 var analytics = require(path.resolve(__dirname, './', 'analytics'));
 var p = new Progger({ speed: 100, token: '.', color: 'blue' });
@@ -147,7 +148,7 @@ module.exports = function (args, opts) {
     var queue = [];
 
     svgData.forEach(function (svg) {
-      return queue.push(util.optimizeSvg(svg));
+      return queue.push(image.optimizeSvg(svg));
     });
 
     return Bluebird.all(queue);
@@ -163,7 +164,7 @@ module.exports = function (args, opts) {
       var svgOut = opts.optimizesvg ? obj.data : obj.original;
       var svgData = {
         svgsrc: svgOut,
-        svgdatauri: util.encodeImage(svgOut, encoding, 'svg'),
+        svgdatauri: image.encode(svgOut, encoding, 'svg'),
         height: util.roundNum(obj.info.height),
         width: util.roundNum(obj.info.width)
       };
@@ -183,7 +184,7 @@ module.exports = function (args, opts) {
 
         _.each(results, function (obj) {
           var destFile = path.resolve(pngDir, obj.name + '.png');
-          return queue.push(util.saveSvgAsPng(obj.svgpath, destFile, obj.height, obj.width));
+          return queue.push(image.saveSvgAsPng(obj.svgpath, destFile, obj.height, obj.width));
         });
 
         // start progress dots
@@ -253,7 +254,7 @@ module.exports = function (args, opts) {
 
         // add to results object
         return _.extend(results[i], {
-          pngdatauri: util.encodeImage(data, 'base64', 'png')
+          pngdatauri: image.encode(data, 'base64', 'png')
         });
       });
     }
